[sw/device/silicon_creator/lib] Prepare to add C warning flags

+++ sw/device/silicon_creator/lib/drivers/otbn_unittest.cc
@@ -219,7 +219,7 @@ TEST_F(DmemWriteTest, FailureOverflowNumWords) {
 TEST_F(DmemWriteTest, FailureOverflowOffset) {
   // Try to trigger an integer overflow with `dest_addr`.
   std::array<uint32_t, 2> test_data = {0x12345678, 0xabcdef01};
-  otbn_addr_t dest_addr = SIZE_MAX;
+  otbn_addr_t dest_addr = UINT32_MAX;
 
   EXPECT_EQ(otbn_dmem_write(test_data.size(), test_data.data(), dest_addr),
             kErrorOtbnBadOffsetLen);
